Updated  sptest() and related plot() function
This handles the upcoming deprecation of sp package. sptest() now requires site location data to be supplied as sf class object, with spatially weights calculated internally. The display function also includes a new option to return sf class object with relevant outputs for user-controlled display. The commit updates vignette and examples, as well as a the help manual.